# Improve performance by avoiding loading the GMT library repeatedly