**Description of proposed changes**

The [`load_libgmt`](https://github.com/GenericMappingTools/pygmt/blob/bb8234530af41ebcfa15ba4d516754ec2c633b45/pygmt/clib/loading.py#L18) function can search for the GMT library and load it so that we can call GMT API functions in PyGMT. The function is called in the [`Session` class](https://github.com/GenericMappingTools/pygmt/blob/bb8234530af41ebcfa15ba4d516754ec2c633b45/pygmt/clib/session.py#L91)'s [`get_libgmt_func`](https://github.com/GenericMappingTools/pygmt/blob/bb8234530af41ebcfa15ba4d516754ec2c633b45/pygmt/clib/session.py#L275) function.

Although the `get_libgmt_func` function is called multiple times in `Session`'s different methods, the GMT library is only loaded once for each session, as shown below:
https://github.com/GenericMappingTools/pygmt/blob/bb8234530af41ebcfa15ba4d516754ec2c633b45/pygmt/clib/session.py#L310-L311

However, when wrapping GMT modules, we always create a new session and then call the GMT module like this
```
with Session() as lib:
    lib.call_module(...)
```
It means the GMT library is searched and loaded multiple times, as reported in #867.

To verify if the GMT library is loaded multiple times, we just need to add a simple print statement like `print("Loading GMT library")` at the start of the `load_libgmt` function, and then call PyGMT. Here is what you'll see. It's clear that the GMT library are loaded multiple times:
```python
In [1]: import pygmt
Loading GMT library
Loading GMT library

In [2]: fig = pygmt.Figure()
Loading GMT library

In [3]: fig.basemap(region=[0, 10, 0, 10], frame=True, projection="X10c")
Loading GMT library
Loading GMT library

In [4]: fig.savefig("map.png")
Loading GMT library
Loading GMT library
```

It's unnecessary to search for and load the GMT library multiple times, which is actually time-consuming (120 ms for me on macOS).
```
In [1]: from pygmt.clib.loading import load_libgmt

In [2]: %timeit load_libgmt()
118 ms ± 5.75 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

This PR fixes the issue by moving the `load_libgmt` call outside the `Session` class, so that the GMT library is only searched for and loaded once when importing pygmt.

This PR addresses the following comment in PR #867 so it fixes #867:

> So it might be time to rethink loading libgmt in the class instead of as a global instead to avoid searching for it every time.

#867 also proposed another alternative solution:

> Or we figure out a way to not call begin at import time.

This solution may help address issue #217, but it's unclear what exactly we should do and it definitely needs big refactors, so we can explore the solution later. 

It's worth noting that the small changes in this PR **speed up our Tests significantly**! For example, the "Run tests" step now only takes **85 seconds** (on Linux) compared to **250 seconds** in the main branch.