@weiji14 2023-12-29
If you click into the link - https://codspeed.io/GenericMappingTools/pygmt/branches/clib/load-libgmt, it shows that most tests improve by 1-2%. Since most unit tests only call a few GMT commands, they shouldn't show too much of a gain, but still, I think it's worth it.

Maybe wait for #2924? There are a few clib/session related tests in that PR which might show a clearer improvement.