@seisman 2023-12-29
> If you click into the link - https://codspeed.io/GenericMappingTools/pygmt/branches/clib/load-libgmt, it shows that most tests improve by 1-2%. Since most unit tests only call a few GMT commands, they shouldn't show too much of a gain, but still, I think it's worth it.
> 
> Maybe wait for #2924? There are a few clib/session related tests in that PR which might show a clearer improvement.

Just tried to benchmark locally. 

In the main branch:
```
In [1]: import pygmt

In [2]: fig = pygmt.Figure()

In [3]: %timeit fig.basemap(region=[0, 10, 0, 10], frame=True, projection="X10c")
237 ms ± 9.54 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
And in this branch:
```
In [1]: import pygmt

In [2]: fig = pygmt.Figure()

In [3]: %timeit fig.basemap(region=[0, 10, 0, 10], frame=True, projection="X10c")
6.12 ms ± 246 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
So, the `Figure.basemap` call is 6.12 ms vs 237 ms. The reduced time (~230 ms) is just the time used to load the GMT library twice. So I expect to see that even the simplest test should have a time reduction of at least 200 ms. But it doesn't match the CodSpeed reports.