@seisman 2023-12-29
I've added a unit test in commit f589ff8b687c9fc35d516175ae731dcf1d64d490. The unit test passes in this branch and fails if in the main branch:

```
_______________________________ TestLibgmtCount.test_libgmt_load_counter _________________________________

self = <pygmt.tests.test_clib_loading.TestLibgmtCount object at 0x7f212fc958b0>

    @pytest.mark.usefixtures("_mock_ctypes")
    def test_libgmt_load_counter(self):
        """
        Make sure that the GMT library is not loaded in every session.
        """
        with Session() as lib:
            _ = lib
        with Session() as lib:
            _ = lib
>       assert self.counter == 0  # ctypes.CDLL is not called after two sessions.
E       assert 2 == 0
E        +  where 2 = <pygmt.tests.test_clib_loading.TestLibgmtCount object at 0x7f212fc958b0>.counter

pygmt/tests/test_clib_loading.py:242: AssertionError

```