@seisman 2023-12-30
> Could we make a unit test based on [#217 (comment)](https://github.com/GenericMappingTools/pygmt/issues/217#issue-344245022), and test to see if the change in this PR allows for this multiprocessing/parallel code to work:

No, it definitely won't work. To allow multiprocessing for PyGMT, we need to call `gmt begin` in each process. Currently, `gmt begin` is only called once when importing pygmt. That's why the trick in https://github.com/GenericMappingTools/pygmt/issues/217#issuecomment-754774875 works.