@seisman 2024-01-01
> > Could we make a unit test based on [#217 (comment)](https://github.com/GenericMappingTools/pygmt/issues/217#issue-344245022), and test to see if the change in this PR allows for this multiprocessing/parallel code to work:
> > ```python
> > import pygmt
> > import multiprocessing as mp
> > 
> > def gmt_func(n):
> >     fig = pygmt.Figure()
> >     fig.coast(...) 
> >     fig.show()
> > 
> > # gmt_func(n=5) #calls the function without multiprocessing - uncomment as desired to see that the function works on its own
> > 
> > with mp.Pool(2) as p: #calls the process with multiprocessing, using 2 cores
> >     a = p.map(gmt_func, [x for x in range(0,2)])
> > ```
> 
> In [48c00ae](https://github.com/GenericMappingTools/pygmt/commit/48c00aee65d9a66ab94770ea6fb7bca6a1ece4d5), I've added a test to make sure that the workaround in #217 still works after this change. But, as mentioned in [#217 (comment)](https://github.com/GenericMappingTools/pygmt/issues/217#issuecomment-762649874), the workaround doesn't work for Windows, so it's marked as `xfail` on Windows.

I've decided to remove this multiprocessing test from this PR, and will add it in PR #2938 and also fix the Windows issue.